Move ApiWorksTests into a separate package
