+++ catalogue_api/api/src/test/scala/uk/ac/wellcome/platform/api/works/ApiWorksTestInvisible.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.api
+package uk.ac.wellcome.platform.api.works
 
 import com.twitter.finagle.http.Status
 import uk.ac.wellcome.models.Work
